+++ src/main/kotlin/org/xbery/artbeams/articles/domain/Article.kt
@@ -4,7 +4,8 @@ import org.xbery.artbeams.common.assets.domain.Asset
 import org.xbery.artbeams.common.assets.domain.AssetAttributes
 import org.xbery.artbeams.common.assets.domain.Validity
 import org.xbery.artbeams.common.assets.domain.ValidityAsset
-import java.util.*
+import java.util.ArrayList
+import java.util.Date
 
 /**
  * Article entity
@@ -19,13 +20,20 @@ data class Article(
     val title: String,
     val image: String?,
     val perex: String,
-    val bodyMarkdown: String,
+    /** Body in editing format (e.g. in markdown). */
+    val bodyEdited: String,
     val body: String,
     val keywords: String,
-    val showOnBlog: Boolean
-) : Asset(), ValidityAsset {
-    fun updatedWith(edited: EditedArticle, htmlBody: String, userId: String): Article {
-        return this.copy(
+    val showOnBlog: Boolean,
+    val editor: String,
+) : Asset(),
+    ValidityAsset {
+    fun updatedWith(
+        edited: EditedArticle,
+        htmlBody: String,
+        userId: String,
+    ): Article =
+        this.copy(
             common = this.common.updatedWith(userId),
             validity = this.validity.updatedWith(edited),
             externalId = (edited.externalId?.let { extId -> if (extId.trim() == "") null else extId }),
@@ -33,23 +41,25 @@ data class Article(
             title = edited.title,
             image = (edited.image?.let { img -> if (img.trim() == "") null else img }),
             perex = edited.perex,
-            bodyMarkdown = edited.bodyMarkdown,
+            bodyEdited = edited.bodyEdited,
             body = htmlBody,
             keywords = edited.keywords,
-            showOnBlog = edited.showOnBlog
+            showOnBlog = edited.showOnBlog,
+            editor = edited.editor,
         )
-    }
 
     fun toEdited(categories: List<String>): EditedArticle {
         val validTo = this.validTo
-        return EditedArticle(this.id,
+        return EditedArticle(
+            this.id,
             this.externalId,
             this.slug,
             this.title,
             this.image,
             null,
             this.perex,
-            this.bodyMarkdown,
+            this.bodyEdited,
+            this.editor,
             if (this.validFrom == AssetAttributes.EMPTY_DATE) {
                 Date()
             } else {
@@ -62,18 +72,27 @@ data class Article(
             },
             this.keywords,
             this.showOnBlog,
-            categories
+            categories,
         )
     }
 
     companion object {
         const val CacheName: String = "articles"
-        val Empty: Article = Article(
-            AssetAttributes.EMPTY,
-            Validity.Empty,
-            null, "", "New article",
-            null, "", "", "", "", true
-        )
+        val Empty: Article =
+            Article(
+                AssetAttributes.EMPTY,
+                Validity.Empty,
+                null,
+                "",
+                "New article",
+                null,
+                "",
+                "",
+                "",
+                "",
+                true,
+                "markdown",
+            )
         val EmptyEdited: EditedArticle = Empty.toEdited(ArrayList<String>())
     }
 }